fix: don't read the cache when serenity potentially could write
Even though serenity could write to the cache at any moment, it's safer to do it before invoking actions that will for sure write to the cache.

You see, for whatever reason, serenity FAILS to write when the cache is read-locked, even if just for an instant to make a copy of a Copy type. Instead of blocking for the miniscule amount of time it would take to get a write lock, which is what a sane library would do, they fail hard and fast.

Supposedly this is fixed in 0.6, but wow.